Rename/relocate some classes and variables
